pyproject.toml
[project]
name = "interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0.0",
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "langchain>=0.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
    "streamlit>=1.55.0",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
]
